revert changes before merge
