+++ src/common/Smi.Common/Options/GlobalOptions.cs
@@ -170,11 +170,32 @@ namespace Smi.Common.Options
         /// </summary>
         public string RedisConnectionString { get; set; }
 
-        public IMappingTableOptions Clone() => (IMappingTableOptions)MemberwiseClone();
-
         public override string ToString()
         {
             return GlobalOptions.GenerateToString(this);
+        }
+
+        public DiscoveredTable Discover()
+        {
+            var server = new DiscoveredServer(MappingConnectionString, MappingDatabaseType);
+
+            var idx = MappingTableName.LastIndexOf('.');
+            var tableNameUnqualified = MappingTableName.Substring(idx + 1);
+
+            idx = MappingTableName.IndexOf('.');
+            if (idx == -1)
+                throw new ArgumentException($"MappingTableName did not contain the database/user section:'{MappingTableName}'");
+
+            var databaseName = server.GetQuerySyntaxHelper().GetRuntimeName(MappingTableName.Substring(0, idx));
+            if (string.IsNullOrWhiteSpace(databaseName))
+                throw new ArgumentException($"Could not get database/username from MappingTableName {MappingTableName}");
+
+            return server.ExpectDatabase(databaseName).ExpectTable(tableNameUnqualified);
+        }
+
+        public IMappingTableOptions Clone()
+        {
+            return (IMappingTableOptions)this.MemberwiseClone();
         }
     }
 
@@ -187,6 +208,7 @@ namespace Smi.Common.Options
         DatabaseType MappingDatabaseType { get; }
         int TimeoutInSeconds { get; }
 
+        DiscoveredTable Discover();
         IMappingTableOptions Clone();
     }
 
@@ -304,7 +326,7 @@ namespace Smi.Common.Options
     }
 
     [UsedImplicitly]
-    public class CohortExtractorOptions : ConsumerOptions, IMappingTableOptions
+    public class CohortExtractorOptions : ConsumerOptions
     {
         private string _auditorType;
 
@@ -318,8 +340,6 @@ namespace Smi.Common.Options
                 : _auditorType;
             set => _auditorType = value;
         }
-
-        public string SwapperType { get; set; }
 
         /// <summary>
         /// The Type of a class implementing IExtractionRequestFulfiller which is responsible for mapping requested image identifiers to image file paths.  Mandatory
@@ -366,17 +386,6 @@ namespace Smi.Common.Options
         /// ID(s) of ColumnInfo that contains a list of values which should not have data extracted for them.  e.g. opt out.  The name of the column referenced must match a column in the extraction table
         /// </summary>
         public List<int> RejectColumnInfos { get; set; }
-
-        public string MappingConnectionString { get; set; }
-
-        public string MappingTableName { get; set; }
-        public string SwapColumnName { get; set; }
-        public string ReplacementColumnName { get; set; }
-
-        public DatabaseType MappingDatabaseType { get; set; }
-
-        public int TimeoutInSeconds { get; set; }
-        public IMappingTableOptions Clone() => (IMappingTableOptions)MemberwiseClone();
 
         public override string ToString()
         {
