--- src/microservices/Microservices.IdentifierMapper/Execution/Swappers/ForGeneratedUIDIdentifierSwapper.cs
@@ -1,9 +1,0 @@
-﻿using Smi.Common.Helpers;
-
-namespace Microservices.IdentifierMapper.Execution.Swappers
-{
-    public class ForGeneratedUIDIdentifierSwapper : ReplacementValueIdentifierSwapper
-    {
-        protected override string GetReplacementValue() => SmiDicomUIDGenerator.Generate();
-    }
-}
