+++ tests/applications/Applications.TriggerUpdates.Tests/MapperSourceTests.cs
@@ -60,7 +60,7 @@ namespace Applications.TriggerUpdates.Tests
                 var swapper = new TableLookupWithGuidFallbackSwapper();
                 swapper.Setup(mapperOptions);
 
-                guidTable = swapper.GetMappingTableIfAny(mapperOptions);
+                guidTable = swapper.GetGuidTableIfAny(mapperOptions);
                 Assert.AreEqual(0,guidTable.GetRowCount(), "No temporary guids should exist yet");
                 Assert.AreEqual(1,map.GetRowCount(),"We should have a mapping table with 1 entry");
